# Pull regex into variables to prevent the regex to compile in the hotpath